Move pagination context functions out of store pkg
With the store methods starting to take in pagination parameters
explicitly, there is no reason for this to live in the store package
anymore.
